feat: add jsii.Sprintf helper
It's a fairly common practice to compose string values using
`fmt.Sprintf`, however in order to use these with `jsii` they need to be
stored into a variable so a pointer can be taken from them, or they need
passed into the `jsii.String` function.

This new helper removes this constraint and provides a simple way to
perform `jsii`-friendly string interpolation.